requirements.txt
# NBA Shot Selection - AISE 4030 Group 11
# Python 3.10+

# Core RL / ML
torch>=2.1.0
torchvision>=0.16.0

# Environment
gymnasium>=0.29.0

# Dataset
datasets>=2.14.0          # HuggingFace datasets (SportVU)
huggingface_hub>=0.19.0

# NBA player data
nba_api>=1.4.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
